Mirror Effect factory in java and scala dsl
