+++ akka-actor-testkit-typed/src/test/scala/akka/actor/testkit/typed/scaladsl/SyncTestingExampleSpec.scala
@@ -7,7 +7,7 @@ package akka.actor.testkit.typed.scaladsl
 //#imports
 import akka.actor.typed._
 import akka.actor.typed.scaladsl._
-import akka.actor.testkit.typed.Effects._
+import akka.actor.testkit.typed.Effect._
 //#imports
 import org.scalatest.{ Matchers, WordSpec }
